First Commit
This is the first commit. So far we have:
1) Created the EtsyDemo
2) Created ability to add, edit and delete new listings
3) Added Bootstrap Functionality
4) Added ability to upload images
5) Created UI functionality through Bootstrap, HTML and CSS
5a) Nav Bar, with custom CSS design
5b) Page Background color
5c) Footer
5d) Buttons
5e) Form design
5f) Grid layout
5g) Learned how to solve for hanging grid elements by sending the 5, 9,
13, etc to the left.
6) Learned how to make partials to simplify code
7) Setup the Homepage to be the same as the Listing Page

I also played around with the Navbar on my own and figured out how to
do the dropdown menu to add the Listings Page, and ability to go to the
Add New listings page so that the link is not hidden at the bottom of
the page like in the class example.

I think that is all for now.